+++ src/styles/index.scss
@@ -13,5 +13,4 @@
   border-right-color: $right-bottom;
   border-style: solid;
   border-top-color: $left-top;
-  border-width: 4px;
 }
