+++ src/utils/index.ts
@@ -1,245 +1,5 @@
-import { MAX_COLS, MAX_ROWS, BOMBS_NUM } from "../constants";
-import { CellState, CellValue, Cell } from "../types";
+import { createField } from "./createField";
+import { grabAllAdjacentCells } from "./grabAdjacentCells";
+import { unwrapCells } from "./unwrapCells";
 
-const grabAllAdjacentCells = (
-  cells: Cell[][],
-  rowParam: number,
-  colParam: number
-): {
-  topLeftCell: Cell | null;
-  topCell: Cell | null;
-  topRightCell: Cell | null;
-  leftCell: Cell | null;
-  rightCell: Cell | null;
-  bottomLeftCell: Cell | null;
-  bottomCell: Cell | null;
-  bottomRightCell: Cell | null;
-} => {
-  const topLeftCell =
-    rowParam > 0 && colParam > 0 ? cells[rowParam - 1][colParam - 1] : null;
-  const topCell = rowParam > 0 ? cells[rowParam - 1][colParam] : null;
-  const topRightCell =
-    rowParam > 0 && colParam < MAX_COLS - 1
-      ? cells[rowParam - 1][colParam + 1]
-      : null;
-  const leftCell = colParam > 0 ? cells[rowParam][colParam - 1] : null;
-  const rightCell =
-    colParam < MAX_COLS - 1 ? cells[rowParam][colParam + 1] : null;
-  const bottomLeftCell =
-    rowParam < MAX_ROWS - 1 && colParam > 0
-      ? cells[rowParam + 1][colParam - 1]
-      : null;
-  const bottomCell =
-    rowParam < MAX_ROWS - 1 ? cells[rowParam + 1][colParam] : null;
-  const bottomRightCell =
-    rowParam < MAX_ROWS - 1 && colParam < MAX_COLS - 1
-      ? cells[rowParam + 1][colParam + 1]
-      : null;
-
-  return {
-    topLeftCell,
-    topCell,
-    topRightCell,
-    leftCell,
-    rightCell,
-    bottomLeftCell,
-    bottomCell,
-    bottomRightCell,
-  };
-};
-
-export const createField = (): Cell[][] => {
-  let field: Cell[][] = [];
-
-  for (let row = 0; row < MAX_ROWS; row++) {
-    field.push([]);
-    for (let col = 0; col < MAX_COLS; col++) {
-      field[row].push({
-        value: CellValue.none,
-        state: CellState.open,
-      });
-    }
-  }
-
-  let bombsPlaced = 0;
-  while (bombsPlaced < BOMBS_NUM) {
-    const y = Math.floor(Math.random() * MAX_ROWS);
-    const x = Math.floor(Math.random() * MAX_COLS);
-
-    if (field[y][x].value !== CellValue.bomb) {
-      field = field.map((row, rowIndex) =>
-        row.map((cell, colIndex) => {
-          if (y === rowIndex && x === colIndex) {
-            return {
-              ...cell,
-              value: CellValue.bomb,
-            };
-          }
-
-          return cell;
-        })
-      );
-      bombsPlaced++;
-    }
-  }
-  for (let rowIndex = 0; rowIndex < MAX_ROWS; rowIndex++) {
-    for (let colIndex = 0; colIndex < MAX_COLS; colIndex++) {
-      const currentCell = field[rowIndex][colIndex];
-      if (currentCell.value === CellValue.bomb) {
-        continue;
-      }
-
-      let numberOfBombs = 0;
-
-      const {
-        topLeftCell,
-        topCell,
-        topRightCell,
-        leftCell,
-        rightCell,
-        bottomLeftCell,
-        bottomCell,
-        bottomRightCell,
-      } = grabAllAdjacentCells(field, rowIndex, colIndex);
-
-      if (topLeftCell?.value === CellValue.bomb) {
-        numberOfBombs++;
-      }
-      if (topCell?.value === CellValue.bomb) {
-        numberOfBombs++;
-      }
-      if (topRightCell?.value === CellValue.bomb) {
-        numberOfBombs++;
-      }
-      if (leftCell?.value === CellValue.bomb) {
-        numberOfBombs++;
-      }
-      if (rightCell?.value === CellValue.bomb) {
-        numberOfBombs++;
-      }
-      if (bottomLeftCell?.value === CellValue.bomb) {
-        numberOfBombs++;
-      }
-      if (bottomCell?.value === CellValue.bomb) {
-        numberOfBombs++;
-      }
-      if (bottomRightCell?.value === CellValue.bomb) {
-        numberOfBombs++;
-      }
-
-      if (numberOfBombs > 0) {
-        field[rowIndex][colIndex] = {
-          ...currentCell,
-          value: numberOfBombs,
-        };
-      }
-    }
-  }
-  return field;
-};
-
-export const unwrapCells = (
-  field: Cell[][],
-  rowParam: number,
-  colParam: number
-): Cell[][] => {
-  let fieldCopy = field.slice();
-  const currentCell = field[rowParam][colParam];
-
-  fieldCopy[rowParam][colParam].state = CellState.visible;
-
-  const {
-    topLeftCell,
-    topCell,
-    topRightCell,
-    leftCell,
-    rightCell,
-    bottomLeftCell,
-    bottomCell,
-    bottomRightCell,
-  } = grabAllAdjacentCells(field, rowParam, colParam);
-
-  if (
-    topLeftCell?.state === CellState.open &&
-    topLeftCell.value !== CellValue.bomb
-  ) {
-    if (topLeftCell.value === CellValue.none) {
-      field = unwrapCells(field, rowParam - 1, colParam - 1);
-    } else {
-      field[rowParam - 1][colParam - 1].state = CellState.visible;
-    }
-  }
-
-  if (topCell?.state === CellState.open && topCell.value !== CellValue.bomb) {
-    if (topCell.value === CellValue.none) {
-      field = unwrapCells(field, rowParam - 1, colParam);
-    } else {
-      field[rowParam - 1][colParam].state = CellState.visible;
-    }
-  }
-
-  if (
-    topRightCell?.state === CellState.open &&
-    topRightCell.value !== CellValue.bomb
-  ) {
-    if (topRightCell.value === CellValue.none) {
-      field = unwrapCells(field, rowParam - 1, colParam + 1);
-    } else {
-      field[rowParam - 1][colParam + 1].state = CellState.visible;
-    }
-  }
-
-  if (leftCell?.state === CellState.open && leftCell.value !== CellValue.bomb) {
-    if (leftCell.value === CellValue.none) {
-      field = unwrapCells(field, rowParam, colParam - 1);
-    } else {
-      field[rowParam][colParam - 1].state = CellState.visible;
-    }
-  }
-
-  if (
-    rightCell?.state === CellState.open &&
-    rightCell.value !== CellValue.bomb
-  ) {
-    if (rightCell.value === CellValue.none) {
-      field = unwrapCells(field, rowParam, colParam + 1);
-    } else {
-      field[rowParam][colParam + 1].state = CellState.visible;
-    }
-  }
-
-  if (
-    bottomLeftCell?.state === CellState.open &&
-    bottomLeftCell.value !== CellValue.bomb
-  ) {
-    if (bottomLeftCell.value === CellValue.none) {
-      field = unwrapCells(field, rowParam + 1, colParam - 1);
-    } else {
-      field[rowParam + 1][colParam - 1].state = CellState.visible;
-    }
-  }
-
-  if (
-    bottomCell?.state === CellState.open &&
-    bottomCell.value !== CellValue.bomb
-  ) {
-    if (bottomCell.value === CellValue.none) {
-      field = unwrapCells(field, rowParam + 1, colParam);
-    } else {
-      field[rowParam + 1][colParam].state = CellState.visible;
-    }
-  }
-
-  if (
-    bottomRightCell?.state === CellState.open &&
-    bottomRightCell.value !== CellValue.bomb
-  ) {
-    if (bottomRightCell.value === CellValue.none) {
-      field = unwrapCells(field, rowParam + 1, colParam + 1);
-    } else {
-      field[rowParam + 1][colParam + 1].state = CellState.visible;
-    }
-  }
-
-  return field;
-};
+export { createField, grabAllAdjacentCells, unwrapCells };
